+++ src/org-pending/org-pending.ts
@@ -4,12 +4,6 @@ import {ModuleWithProviders} from '@angular/core';
 import { PendingOverviewComponent } from 'src/org-pending/containers/overview/pending-overview.component'
 
 export const ROUTES: Routes = [
-  /*{
-    path: 'pending',
-    component: OrgPendingComponent,
-    canActivate: [
-    ],
-  }*/
   {
     path: 'pending',
     component: PendingOverviewComponent,
@@ -17,6 +11,5 @@ export const ROUTES: Routes = [
     ],
   }
 ];
-
 
 export const orgPendingRouting: ModuleWithProviders = RouterModule.forChild(ROUTES);
