+++ src/org-pending/state/actions/index.ts
@@ -1,1 +1,1 @@
-export * from './org-pending.actions';
+export * from './pending-org.actions';
